Optimize database deserialization to better use doc IDs
We had two unbounded arrays in SearchDefinition and Search.
Per https://docs.atlas.mongodb.com/schema-suggestions/avoid-unbounded-arrays/ and PouchDB docs,
this is discouraged since each element added to the array would require a new revision and may
conflict with other changes. See points 6 and 7 of PouchDB best pracices:

https://pouchdb.com/2014/06/17/12-pro-tips-for-better-code-with-pouchdb.html#:~:text=6.%20Don't%20just%20update%20docs%20for%20the%20hell%20of%20it

We fix this by using the allDocs index and the keys/subkeys that we set up to
pull the data we're looking for.

There's a bit of a headache because the query for a single-level subkey will match all further subkey levels
(i.e. key/sub also matches key/sub/sub2/sub3), but we can handle that alright for now.